@@ -4,6 +4,7 @@
   "description": "A simple, accessible and customizable HTML5, YouTube and Vimeo media player",
   "homepage": "https://plyr.io",
   "author": "Sam Potts <sam@potts.es>",
+  "type": "module",
   "main": "dist/plyr.js",
   "types": "src/js/plyr.d.ts",
   "module": "dist/plyr.min.mjs",
@@ -40,64 +41,63 @@
     "start": "gulp"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.2",
+    "@babel/core": "^7.28.3",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-optional-chaining": "^7.18.9",
-    "@babel/plugin-transform-nullish-coalescing-operator": "^7.24.1",
-    "@babel/preset-env": "^7.20.2",
+    "@babel/plugin-proposal-optional-chaining": "^7.21.0",
+    "@babel/plugin-transform-nullish-coalescing-operator": "^7.27.1",
+    "@babel/preset-env": "^7.28.3",
     "@rollup/plugin-babel": "^6.0.4",
-    "@sampotts/eslint-config": "1.1.7",
-    "autoprefixer": "^10.4.13",
-    "aws-sdk": "^2.1256.0",
+    "@rollup/plugin-commonjs": "^28.0.6",
+    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@sampotts/eslint-config": "2.0.3",
+    "autoprefixer": "^10.4.21",
+    "aws-sdk": "^2.1692.0",
     "babel-eslint": "^10.1.0",
-    "browser-sync": "^2.27.10",
-    "colorette": "2.0.19",
-    "cspell": "^6.14.2",
-    "cssnano": "^5.1.14",
-    "del": "^6.1.1",
-    "eslint": "^7.23.0",
+    "browser-sync": "^3.0.4",
+    "colorette": "2.0.20",
+    "cspell": "^9.2.0",
+    "cssnano": "^7.1.1",
+    "del": "^8.0.0",
+    "eslint": "^8.57.1",
     "fancy-log": "^2.0.0",
     "git-branch": "^2.0.1",
-    "gulp": "^4.0.2",
-    "gulp-awspublish": "^6.0.2",
+    "gulp": "^5.0.1",
+    "gulp-awspublish": "^8.0.0",
     "gulp-better-rollup": "^4.0.1",
-    "gulp-filter": "^7.0.0",
+    "gulp-filter": "^9.0.1",
     "gulp-header": "^2.0.9",
-    "gulp-hub": "^4.2.0",
     "gulp-if": "^3.0.0",
-    "gulp-imagemin": "^7.1.0",
+    "gulp-imagemin": "^9.1.0",
     "gulp-open": "^3.0.1",
     "gulp-plumber": "^1.2.1",
-    "gulp-postcss": "^9.0.1",
-    "gulp-rename": "^2.0.0",
-    "gulp-replace": "^1.1.3",
-    "gulp-sass": "^5.1.0",
-    "gulp-size": "^4.0.1",
+    "gulp-postcss": "^10.0.0",
+    "gulp-rename": "^2.1.0",
+    "gulp-replace": "^1.1.4",
+    "gulp-sass": "^6.0.1",
+    "gulp-size": "^5.0.0",
     "gulp-sourcemaps": "^3.0.0",
     "gulp-svgstore": "^9.0.0",
     "gulp-terser": "^2.1.0",
-    "postcss": "^8.4.19",
-    "postcss-custom-properties": "^12.1.9",
-    "postcss-scss": "^4.0.5",
-    "prettier-eslint": "^12.0.0",
+    "imagemin-svgo": "^11.0.1",
+    "postcss": "^8.5.6",
+    "postcss-custom-properties": "^14.0.6",
+    "postcss-scss": "^4.0.9",
+    "prettier-eslint": "^16.4.2",
     "prettier-stylelint": "^0.4.2",
-    "remark-cli": "^11.0.0",
-    "remark-validate-links": "^12.1.0",
-    "rollup": "^3.3.0",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "sass": "^1.56.1",
-    "stylelint": "^14.15.0",
-    "stylelint-config-prettier": "^9.0.4",
-    "stylelint-config-sass-guidelines": "^9.0.1",
-    "stylelint-selector-bem-pattern": "^2.1.1"
+    "remark-cli": "^12.0.1",
+    "remark-validate-links": "^13.1.0",
+    "rollup": "^4.48.1",
+    "sass": "^1.91.0",
+    "stylelint": "^16.23.1",
+    "stylelint-config-prettier": "^9.0.5",
+    "stylelint-config-sass-guidelines": "^12.1.0",
+    "stylelint-selector-bem-pattern": "^4.0.1"
   },
   "dependencies": {
-    "core-js": "^3.26.1",
+    "core-js": "^3.45.1",
     "custom-event-polyfill": "^1.0.7",
-    "loadjs": "^4.2.0",
+    "loadjs": "^4.3.0",
     "rangetouch": "^2.0.1",
-    "url-polyfill": "^1.1.12"
+    "url-polyfill": "^1.1.13"
   }
 }
